improve ui

+++ db/migrate/20240208101929_create_join_table_person_exam.rb
@@ -3,7 +3,7 @@ class CreateJoinTablePersonExam < ActiveRecord::Migration[7.1]
     create_table :people_has_exams do |t|
       t.belongs_to :person, foreign_key: true
       t.belongs_to :exam, foreign_key: true
-      t.float :grade
+      t.decimal :grade
     end
   end
 end
